+++ SEP490_G41/DataAccess/DAO/MappointDAO.cs
@@ -12,17 +12,30 @@ namespace DataAccess.DAO
 {
     public class MappointDAO
     {
-        private readonly finsContext _context; private readonly WKBReader _wkbReader;
+        private readonly finsContext _context;
 
         public MappointDAO(finsContext context)
         {
             _context = context;
-            _wkbReader = new WKBReader();
+        }
+
+        public MappointDAO()
+        {
+            _context = new finsContext();
         }
 
         // Thêm mới mappoint
         public void AddMappoint(Mappoint mappoint)
         {
+            if (mappoint == null)
+                throw new ArgumentNullException(nameof(mappoint));
+
+            if (mappoint.MapId <= 0)
+                throw new ArgumentException("Map ID must be a positive integer.");
+
+            if (mappoint.Location == null)
+                throw new ArgumentException("Location cannot be null.");
+
             _context.Mappoints.Add(mappoint);
             _context.SaveChanges();
         }
@@ -30,68 +43,70 @@ namespace DataAccess.DAO
         // Đọc thông tin mappoint bằng Id
         public Mappoint GetMappointById(int mappointId)
         {
+            if (mappointId <= 0)
+                throw new ArgumentException("Mappoint ID must be a positive integer.");
+
             return _context.Mappoints.FirstOrDefault(mp => mp.MapPointId == mappointId);
         }
 
         // Cập nhật thông tin mappoint
         public void UpdateMappoint(Mappoint mappoint)
         {
-            var existingMappoint = _context.Mappoints.FirstOrDefault(mp => mp.MapPointId == mappoint.MapPointId);
+            if (mappoint == null)
+                throw new ArgumentNullException(nameof(mappoint));
 
+            if (mappoint.MapPointId <= 0)
+                throw new ArgumentException("Mappoint ID must be a positive integer.");
+
+            if (mappoint.MapId <= 0)
+                throw new ArgumentException("Map ID must be a positive integer.");
+
+            if (mappoint.Location == null)
+                throw new ArgumentException("Location cannot be null.");
+
+            var existingMappoint = _context.Mappoints.FirstOrDefault(mp => mp.MapPointId == mappoint.MapPointId);
             if (existingMappoint != null)
             {
                 existingMappoint.MapId = mappoint.MapId;
                 existingMappoint.Location = mappoint.Location;
-                // Cập nhật các thuộc tính khác nếu cần
-
                 _context.SaveChanges();
+            }
+            else
+            {
+                throw new ArgumentException($"Mappoint with ID {mappoint.MapPointId} does not exist.");
             }
         }
 
         // Xóa mappoint bằng Id
+       
         public void DeleteMappoint(int mappointId)
         {
-            var mappoint = _context.Mappoints.FirstOrDefault(mp => mp.MapPointId == mappointId);
+            if (mappointId <= 0)
+                throw new ArgumentException("Mappoint ID must be a positive integer.");
+
+            var mappoint = _context.Mappoints.Find(mappointId);
             if (mappoint != null)
             {
                 _context.Mappoints.Remove(mappoint);
                 _context.SaveChanges();
             }
+            else
+            {
+                throw new ArgumentException($"Mappoint with ID {mappointId} does not exist.");
+            }
         }
+
 
         public List<Mappoint> GetAllMappoints()
         {
             try
             {
-                var mappoints = _context.Mappoints.Include(mp => mp.Map).ToList();
-
-
-                /*// Đọc dữ liệu địa lý từ MySQL dưới dạng chuỗi và chuyển đổi thành kiểu NetTopologySuite.Point
-                foreach (var mappoint in mappoints)
-                {
-                    // Đọc dữ liệu POINT từ MySQL dưới dạng chuỗi
-                    string locationString = mappoint.Location.ToString();
-                    Console.WriteLine("LocationString: " + locationString);
-
-                    // Chuyển đổi chuỗi thành đối tượng Point của NetTopologySuite
-                    var geometryFactory = NtsGeometryServices.Instance.CreateGeometryFactory(srid: 4326);
-                    var wktReader = new WKTReader(geometryFactory);
-                    var point = (Point)wktReader.Read(locationString);
-
-                    // Gán đối tượng Point cho thuộc tính Location của đối tượng Mappoint
-                    mappoint.Location = point;
-                }*/
-
-                return mappoints;
+                return _context.Mappoints.Include(mp => mp.Map).ToList();
             }
             catch (Exception ex)
             {
                 throw new Exception("Error occurred while getting all map points.", ex);
             }
         }
-
-
     }
-
 }
-
